So this works for 12 but not 13
I was making an assumptions about padding sizes which I would rather
not have to make. If we *assume* that we are inserting into the first
block of content then our technique from 12 works. However, if we are
*not* inserting into the first block then our technique doesn't work at
all. In this case, it will think it's found a key size that looks like
some multiple of the actual key size (depending on how much content
there is before wherever our padding is going, e.g. if we are inserting
into the second block it will think the key size is 32, the third block
48, etc. etc.).

I'm refactoring this to recognize that we only know the key size when we
have seen two things:
1) a block stops changing *that was previously changing* (note: when
this happens we will know how offset we are from the end of the block
that we are inserting into, and we will know where that block ends).
2) the block *after* the one we are inserting into stops changing (at
this point our padding will be the length of our offset from up above
plus the block size).